device && patient manage

+++ rdexpense-manager/src/main/java/com/rdexpense/manager/controller/imgSys/DeviceController.java
@@ -11,10 +11,13 @@ import com.github.pagehelper.PageInfo;
 import com.rdexpense.manager.controller.base.BaseController;
 import com.rdexpense.manager.dto.imgSys.BodyPart;
 import com.rdexpense.manager.dto.imgSys.Device;
+import com.rdexpense.manager.dto.imgSys.Page.PageParam;
+import com.rdexpense.manager.dto.imgSys.serialNum.SerialNumListDto;
 import com.rdexpense.manager.service.imgSys.DeviceService;
 import com.rdexpense.manager.util.LogUtil;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -53,6 +56,7 @@ public class DeviceController extends BaseController {
         CheckParameter.stringLengthAndEmpty(pd.getString("deviceSerialNum"), "设备序列号(Mac地址)",128);
         CheckParameter.stringLengthAndEmpty(pd.getString("deviceName"), "设备名称",128);
         CheckParameter.stringLengthAndEmpty(pd.getString("deviceCode"), "设备资产编码",128);
+        CheckParameter.stringLengthAndEmpty(pd.getString("deviceModel"), "设备型号",128);
         CheckParameter.stringLengthAndEmpty(pd.getString("departmentCode"), "所属科室编码",128);
 
         ResponseEntity result = null;
@@ -69,11 +73,14 @@ public class DeviceController extends BaseController {
         }
     }
 
-    @ApiOperation(value = "删除某设备信息")
+    @ApiOperation(value = "删除一个/多个设备信息")
     @DeleteMapping()
-    public ResponseEntity deleteDevice(Device device) {
+    public ResponseEntity deleteDevice(SerialNumListDto serialNumListDto) {
         PageData pd = this.getParams();
-        CheckParameter.stringLengthAndEmpty(pd.getString("deviceSerialNum"), "设备序列号",128);
+        String idList = pd.getString("idList");
+        if (StringUtils.isBlank(idList)) {
+            throw new MyException("序列号不能为空!");
+        }
         ResponseEntity result = null;
         try {
             deviceService.deleteDevice(pd);
@@ -118,17 +125,29 @@ public class DeviceController extends BaseController {
      */
     @GetMapping()
     @ApiOperation(value = "分页(关键词)查询所有设备信息")
-    public ResponseEntity<PageInfo<Device>> getAllDevice(Device device) {
+    public ResponseEntity<PageInfo<Device>> getAllDevice(Device device, PageParam pageParam) {
         PageData pd = this.getParams();
-//        CheckParameter.stringLengthAndEmpty(pd.getString("pageNum"), "查询页码",128);
-//        CheckParameter.stringLengthAndEmpty(pd.getString("pageSize"), "每页大小",128);
         try {
+            // 这里方便起见先转String，后真正分页时将数据再转回Int即可
             pd.put("pageNum", pd.getString("pageNum") == "" ? "1" : pd.getString("pageNum"));
             pd.put("pageSize", pd.getString("pageSize") == "" ? "10" : pd.getString("pageSize"));
             PageHelper.startPage(pd.getInt("pageNum"), pd.getInt("pageSize"));
             List<PageData> list = deviceService.getAllDevice(pd);
             PageInfo<PageData> pageInfo = new PageInfo<>(list);
             return ResponseEntity.success(PropertyUtil.pushPageList(pageInfo, Device.class), ConstantMsgUtil.INFO_QUERY_SUCCESS.desc());
+        } catch (MyException e) {
+            logger.error("查询所有设备信息,request=[{}]", pd);
+            return ResponseEntity.failure(ERR_QUERY_FAIL.val(), ERR_QUERY_FAIL.desc());
+        }
+    }
+
+    @GetMapping("/sketch")
+    @ApiOperation(value = "新增设备/根据设备id查找部位时的所有部位以及相应示意图")
+    public ResponseEntity<List<BodyPart>> getBodyPartByDevice(Device device) {
+        PageData pd = this.getParamsFormat("XML");
+        try {
+            List<PageData> list = deviceService.getBodyPartByDevice(pd);
+            return ResponseEntity.success(PropertyUtil.covertListModel(list, BodyPart.class), ConstantMsgUtil.INFO_QUERY_SUCCESS.desc());
         } catch (MyException e) {
             logger.error("查询所有设备信息,request=[{}]", pd);
             return ResponseEntity.failure(ERR_QUERY_FAIL.val(), ERR_QUERY_FAIL.desc());
